+++ components/comp_adc/include/comp_adc.h
@@ -1,2 +1,1 @@
-void func(void);
-void vTaskADC(void *vparameter);
+void vTaskADC(void *pvParameters);
