+++ components/comp_blink/include/comp_blink.h
@@ -1,1 +1,1 @@
-void func(void);
+void vTaskBlink(void *pvParameters);
