+++ components/comp_input/include/comp_input.h
@@ -1,1 +1,1 @@
-void func(void);
+void vTaskInput(void *pvParameters);
